Move makeUniqueId to internal
