+++ src/utilities/internal/index.ts
@@ -31,6 +31,7 @@ export { isNonEmptyArray } from "./isNonEmptyArray.js";
 export { isNonNullObject } from "./isNonNullObject.js";
 export { isPlainObject } from "./isPlainObject.js";
 export { makeReference } from "./makeReference.js";
+export { makeUniqueId } from "./makeUniqueId.js";
 export { maybeDeepFreeze } from "./maybeDeepFreeze.js";
 export { mergeDeep } from "./mergeDeep.js";
 export { mergeDeepArray } from "./mergeDeepArray.js";
